requirements.txt
fastapi
uvicorn
jinja2>=3.1
python-multipart>=0.0.9

SQLAlchemy>=2.0
asyncpg>=0.29
alembic>=1.13

psycopg2-binary>=2.9
psycopg2

pydantic>=2.6
pydantic-settings>=2.2

strawberry-graphql[fastapi]>=0.230
requests>=2.31
email-validator>=2.1

httpx>=0.27.0
pyserial>=3.5
pyserial-asyncio>=0.6